+++ packages/components/src/__overlays__/Tooltip/v3/_docs/Tooltip.mdx
@@ -1,12 +1,13 @@
 import { Canvas, Meta, Unstyled, Controls } from "@storybook/blocks"
 import { CheckIcon, ClearIcon } from "~components/Icon"
 import { ResourceLinks, SbText, SbHeading, Installation } from "~storybook/components"
-import * as TooltipStories from "./Tooltip.stories"
+import * as TooltipStories from "./Tooltip.docs.stories"
+import * as ExampleStories from "./Tooltip.stories"
 import AnatomyPng from "./assets/tooltip_anatomy.png"
 import VariantsPng from "./assets/tooltip_variants.png"
 import WhenToUsePng from "./assets/tooltip_when_to_use.png"
 
-<Meta title="Overlays/Tooltip/v3/Usage Guidelines" />
+<Meta title="Overlays/Tooltip/v3/Usage Guidelines" of={ExampleStories} />
 
 # Tooltip (v3)
 <SbText variant="extra-small">
